User now receives updates on overspending
Lots of features still need to be updated. The user can now view their inbox and see if they have spent more than they made, and also when they spend more than the "rule" they have made for themselves. Will add in a feature to not allow duplicate messages, and allow the user to actually read the message rather than it being hardcoded as "unread"