Added a button to read messages
Added button functionality and (un)bolding to show when a message is read or not. Also updated the transactions to correctly reflect income vs. expenses. Allowing the user to accurately view how much they are spending compared to how much they are being paid (also prints to console for debug purposes)