Added in filter by date
Added in an option to filter by day, week, month, year, and overall transactions with a reflecting pie chart. Also updated the csv to have various dates in order to test the program properly. I had to get rid of the cats because the site reloads every time you change the period (sorry)